\name{degree_hour_limits}
\alias{degree_hour_limits}
\title{Calculates temperature limits for plots.}
\usage{
degree_hour_limits(balance_point = BALANCE_POINT)
}
\arguments{
  \item{balance_point}{the balance point to use in
  calculating the high and low limits of expected degree
  hours}
}
\description{
Calculates temperature limits for plots.
}

