\docType{data}
\name{forecast.dh}
\alias{forecast.dh}
\title{Degree-hour forecast data calculated from the temperature forcasts on March 19 2014}
\description{
Degree-hour forecast data calculated from the temperature
forcasts on March 19 2014
}
\author{
Yoni Ben-Meshulam \email{yoni.bmesh@gmail.com}
}
\keyword{data}

