\docType{data}
\name{forecast}
\alias{forecast}
\title{Raw forecast data taken from the National Weather Forecast Database March 19 2014}
\description{
Raw forecast data taken from the National Weather Forecast
Database March 19 2014
}
\author{
Yoni Ben-Meshulam \email{yoni.bmesh@gmail.com}
}
\keyword{data}

