\docType{data}
\name{HIGH_TEMPERATURE}
\alias{HIGH_TEMPERATURE}
\title{Highest north american temperature extreme in degrees Celsius.
See http://www.ncdc.noaa.gov/oa/climate/globalextremes.html}
\format{\preformatted{ num 56.7
}}
\usage{
HIGH_TEMPERATURE
}
\description{
Highest north american temperature extreme in degrees
Celsius. See
http://www.ncdc.noaa.gov/oa/climate/globalextremes.html
}
\keyword{datasets}

