\docType{data}
\name{LOW_TEMPERATURE}
\alias{LOW_TEMPERATURE}
\title{Lowest north american temperature extreme in degrees Celsius.
See http://www.ncdc.noaa.gov/oa/climate/globalextremes.html}
\format{\preformatted{ num -63
}}
\usage{
LOW_TEMPERATURE
}
\description{
Lowest north american temperature extreme in degrees
Celsius. See
http://www.ncdc.noaa.gov/oa/climate/globalextremes.html
}
\keyword{datasets}

