\name{plot_color_bar}
\alias{plot_color_bar}
\title{Plots a color bar, to be used as a legend.}
\usage{
plot_color_bar(balance_point = BALANCE_POINT)
}
\arguments{
  \item{balance_point}{the balance point to use in
  calculating the color bar limits}
}
\description{
Plots a color bar, to be used as a legend.
}
\examples{
plot_color_bar()
}

