\name{plot_dh}
\alias{plot_dh}
\title{Plots the degree-hours.}
\usage{
plot_dh(forecast, attribute, balance_point = BALANCE_POINT)
}
\arguments{
  \item{forecast}{the weather forecast}

  \item{attribute}{the attribute to plot}
}
\description{
Plots the degree-hours.
}
\examples{
data(forecast.dh)
}

